// src/enums/boletoType.ts
export enum BoletoType {
  BANK = 'BANK',
  COLLECTION_GOVERNMENT_AGENCIES = 'COLLECTION_GOVERNMENT_AGENCIES',
  COLLECTION_TOWN_HALL = 'COLLECTION_TOWN_HALL',
  COLLECTION_TRANSIT_FEE = 'COLLECTION_TRANSIT_FEE',
  COVENANT_ELECTRICITY_AND_GAS = 'COVENANT_ELECTRICITY_AND_GAS',
  COVENANT_SANITATION = 'COVENANT_SANITATION',
  COVENANT_TELECOMMUNICATION = 'COVENANT_TELECOMMUNICATION',
  CREDIT_CARD = 'CREDIT_CARD',
  OTHERS = 'OTHERS',
}